css/css3-selectors/i18n/css3-selectors-lang-046.html

<!DOCTYPE html>
<html  lang="en" >
<head>
<meta charset="utf-8"/>
<title>[lang="es-MX"], lang="es"</title>
<link rel='author' title='Richard Ishida' href='mailto:ishida@w3.org'>
<link rel='help' href='http://www.w3.org/TR/css3-selectors/#attribute-selectors'>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style type='text/css'>
.test div { width: 100px; height: 50px; background-color: blue; }
#box[lang='es-MX'] { width: 50px; }
.ref { width: 100px; height: 50px; background-color: blue; }
</style>
</head>
<body>


<p>The following two blocks should be <strong>identical</strong>.</p>
<div class="test"><div id="box" lang="es">&#xA0;</div></div>
<p class="ref">&#xA0;</p>


<script>
test(function() {
assert_equals(document.getElementById('box').offsetWidth, 100);
}, "A lang= value will NOT match a lang attribute value when the former contains more subtags.");
</script>

<div id='log'></div>

</body>
</html>
